pyproject.toml
[tool.poetry]
name = "newsparser"
version = "0.1.0"
description = "Project for parsing news from different sources and save them in database and notify in telegram"
authors = ["alexander <as.frantsev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
feedparser = "^6.0.10"
telethon = "^1.29.2"
pymongo = "^4.4.1"
loguru = "^0.7.0"
beautifulsoup4 = "^4.12.2"
python-dotenv-vault = "^0.6.3"
scrapy = "^2.10.0"
python-telegram-bot = {extras = ["job-queue"], version = "^20.4"}
aiohttp = "^3.8.5"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"


[tool.poetry.group.web.dependencies]
flask = "^2.3.2"
pymongo = "^4.4.1"
python-dotenv-vault = "^0.6.3"
gunicorn = "^21.2.0"
eventlet = "^0.33.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
